Refactor instruction module
Based on comments in a previous PR, where the `Instruction` trait seemed convoluted, I decided to replace it by an `enum`. Now, instead of forcing each instruction to conform to the trait, it's the implementation of the `Instruction` enum that is in charge of creating and executing instructions based on it's variant. The `vm` module now only needs to deal with the `Instruction` enum for creating and executing instructions.

These changes improve the understadability of the codebase, by having the matching explicitly in the `Instruction` implementation. It also frees each instruction to implement a custom interface (by not having to conform to the `Instruction` trait). On the downside, what before was a single match, is now two matches where each instruction must be added.